blendy hops to the left

+++ src/world_system.hpp
@@ -83,7 +83,7 @@ private:
 	void handlePlayerMovement(int key, int action);
 	void update_player_movement();
 	void move_player(vec2 direction);
-	void get_blendy_render_request(bool up, bool down, bool left, bool right, float stage);
+	void get_blendy_render_request(bool up, bool down, bool left, bool right, int stage);
 };
